Refactor complete Signed-off-by: Joey Smith <jsmith@webinertia.net>
